make HlsServerService
